+++ tests/integration/test_startup_scripts/configs/bad_script.xml
