+++ tests/integration/test_startup_scripts/configs/good_script.xml
