+++ tests/integration/test_startup_scripts/configs/query_log.xml
