+++ tests/integration/test_startup_scripts/configs/startup_scripts.xml
