--- tests/integration/test_startup_scripts/configs/users.xml
@@ -1,41 +1,0 @@
-<clickhouse>
-
-    <profiles>
-        <default>
-        </default>
-
-        <readonly>
-            <readonly>1</readonly>
-        </readonly>
-    </profiles>
-
-    <users>
-        <default>
-
-            <password></password>
-
-            <networks>
-                <ip>::/0</ip>
-            </networks>
-
-            <profile>default</profile>
-
-            <quota>default</quota>
-
-        </default>
-    </users>
-
-    <quotas>
-        <default>
-            <interval>
-                <duration>3600</duration>
-
-                <queries>0</queries>
-                <errors>0</errors>
-                <result_rows>0</result_rows>
-                <read_rows>0</read_rows>
-                <execution_time>0</execution_time>
-            </interval>
-        </default>
-    </quotas>
-</clickhouse>
