--- tests/integration/test_startup_scripts_execution_state/config/users.xml
@@ -1,8 +1,0 @@
-<clickhouse>
-    <users>
-        <default>
-            <profile>default</profile>
-            <no_password></no_password>
-        </default>
-    </users>
-</clickhouse>
